Turn all exponents into superscript
We already use unicode, so why not use superscript exponents.
The formatting of Arrow is done in rust-simplicity, so we cannot make
any changes for now. We could update how scribe types are printed, but
this would be inconsistent with Arrow. Instead, we find all
occurrences of the 2^[0-9]+ pattern with a regex and turn the exponent
into superscript. If superscripts are popular enough, we can add this to
rust-simplicity directly.